Restructured commands.
