+++ BotInfrastructure/Abstract/RegexCommand.cs
@@ -1,4 +1,4 @@
-﻿namespace BotInfrastructure.Interface.Command;
+﻿namespace BotInfrastructure.Abstract;
 
 public abstract class RegexCommand(string name, string validRegexArg) : MultipleRegexCommand(name, [validRegexArg])
 {
